+++ util/size.go
@@ -3,9 +3,9 @@ package util
 import "unsafe"
 
 const (
-	SizeOfString  = uint64(unsafe.Sizeof(*new(string)))
-	SizeOfUint32  = uint64(unsafe.Sizeof(*new(uint32)))
-	SizeOfUint64  = uint64(unsafe.Sizeof(*new(uint64)))
-	SizeOfPointer = uint64(unsafe.Sizeof(new(int)))
-	SizeOfFloat64 = uint64(unsafe.Sizeof(*new(float64)))
+	SizeOfString  = int(unsafe.Sizeof(*new(string)))
+	SizeOfUint32  = int(unsafe.Sizeof(*new(uint32)))
+	SizeOfUint64  = int(unsafe.Sizeof(*new(uint64)))
+	SizeOfPointer = int(unsafe.Sizeof(new(int)))
+	SizeOfFloat64 = int(unsafe.Sizeof(*new(float64)))
 )
